squashme: fix circular imports
squashme: revert uneeded change
