+++ components/renku_data_services/k8s_watcher/__init__.py
@@ -2,5 +2,6 @@
 
 from renku_data_services.k8s_watcher.core import K8sWatcher, k8s_object_handler
 from renku_data_services.k8s_watcher.db import K8sDbCache
+from renku_data_services.k8s_watcher.orm import BaseORM
 
-__all__ = ["K8sWatcher", "k8s_object_handler", "K8sDbCache"]
+__all__ = ["K8sWatcher", "k8s_object_handler", "K8sDbCache", "BaseORM"]
